src/slam6d/transformFrames.cc: use int instead of ssize_t to make MSVC happy
git-svn-id: https://svn.code.sf.net/p/slam6d/code/trunk@1057 51f1e6e7-7289-49ad-8748-f09e43fc6daa

+++ src/slam6d/transformFrames.cc
@@ -276,7 +276,7 @@ void modifyFrames(double* resTrans, string &dir, int start, int end, bool inputI
     ofstream modFrames_out;
     ofstream modFrames_out2;
 
-    for (ssize_t i = start; i <= end; i++){
+    for (int i = start; i <= end; i++){
         snprintf(inputFileName, 255, "%sscan%.3zd.frames", dir.c_str(), i);
         snprintf(modFramesFileName, 255, "%sMODIFIED_scan%.3zd.frames", dir.c_str(), i);
